Better integration of the hatari emulator
The way hatari emulator is provided has been improved. It is now executed
in real full screen and there is now no tearing at all. The resolution
switch is not made by the wrapper before starting emulator anymore but by
the emulator itself through the use of a custom modeline defined on
xorg.conf. The documentation has been updated with steps to configure it.